# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - email
    - id
    - name
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  handler.HealthResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      status:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A CRUD REST API built with Gin, PostgreSQL, Docker, and Hexagonal Architecture.
  title: Hexagonal Architecture CRUD API
  version: "1.0"
paths:
  /api/v1/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image to the object storage.
      parameters:
      - description: Image to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      summary: Upload an image
      tags:
      - Files
  /api/v1/users:
    get:
      description: Get all users.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user.
      parameters:
      - description: Create user payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
      summary: Create user
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      description: Delete an existing user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      summary: Delete user
      tags:
      - Users
    get:
      description: Get one user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update an existing user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      summary: Update user
      tags:
      - Users
  /health:
    get:
      description: Check API health status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HealthResponse'
      summary: Health check
      tags:
      - Health
swagger: "2.0"